feat(member_join): Added a keypad with buttons to help prevent against bots
The goal of this is mostly to prevent bots. Upon joining, members press a button and that will give them a keypad of buttons for which to enter the code. If the code is right then itll grant them the role but if its wrong itll moderators of so
we went through some iterations of this and ended up with this one. Currently, bots arent able to use interactions so it works